move lib and add new next process system
