+++ souls/example-twenty-questions/soul/cognitiveSteps/brainstorm.ts
