+++ souls/example-twenty-questions/soul/cognitiveSteps/decision.ts
