+++ souls/example-twenty-questions/soul/cognitiveSteps/externalDialog.ts
