+++ souls/example-twenty-questions/soul/cognitiveSteps/mentalQuery.ts
