+++ souls/example-twenty-questions/soul/mentalProcesses/answersQuestions.ts
@@ -1,7 +1,7 @@
 import { MentalProcess, useProcessMemory, useActions } from "@opensouls/engine";
-import brainstorm from "../lib/brainstorm.js";
-import externalDialog from "../lib/externalDialog.js";
-import mentalQuery from "../lib/mentalQuery.js";
+import brainstorm from "../cognitiveSteps/brainstorm.js";
+import externalDialog from "../cognitiveSteps/externalDialog.js";
+import mentalQuery from "../cognitiveSteps/mentalQuery.js";
 
 const answersGuesses: MentalProcess<{object: string}> = async ({ workingMemory, params: { object } }) => {
   const questionsAttempted = useProcessMemory(0);
